fixup! fixup! fixup! fixup! fixup! Minipod only runs every 6 cycles

+++ src/oscillator-disciplining.c
@@ -266,14 +266,14 @@ static float filter_phase(struct kalman_parameters *kalman, float phase, int int
 static void print_inputs(struct od_input inputs[7])
 {
 	log_info(
-		"Inputs: [%ld, %ld, %ld, %ld, %ld, %ld, %ld]",
-		inputs[0].phase_error.tv_nsec+ (float) inputs[0].qErr / PS_IN_NS,
-		inputs[1].phase_error.tv_nsec+ (float) inputs[1].qErr / PS_IN_NS,
-		inputs[2].phase_error.tv_nsec+ (float) inputs[2].qErr / PS_IN_NS,
-		inputs[3].phase_error.tv_nsec+ (float) inputs[3].qErr / PS_IN_NS,
-		inputs[4].phase_error.tv_nsec+ (float) inputs[4].qErr / PS_IN_NS,
-		inputs[5].phase_error.tv_nsec+ (float) inputs[5].qErr / PS_IN_NS,
-		inputs[6].phase_error.tv_nsec+ (float) inputs[6].qErr / PS_IN_NS
+		"Inputs: [%f, %f, %f, %f, %f, %f, %f]",
+		(float) inputs[0].phase_error.tv_nsec + (float) inputs[0].qErr / PS_IN_NS,
+		(float) inputs[1].phase_error.tv_nsec + (float) inputs[1].qErr / PS_IN_NS,
+		(float) inputs[2].phase_error.tv_nsec + (float) inputs[2].qErr / PS_IN_NS,
+		(float) inputs[3].phase_error.tv_nsec + (float) inputs[3].qErr / PS_IN_NS,
+		(float) inputs[4].phase_error.tv_nsec + (float) inputs[4].qErr / PS_IN_NS,
+		(float) inputs[5].phase_error.tv_nsec + (float) inputs[5].qErr / PS_IN_NS,
+		(float) inputs[6].phase_error.tv_nsec + (float) inputs[6].qErr / PS_IN_NS
 	);
 }
